Fix: update inputs and test assertions for decimal money conversion
- Changed `long UnitPrice` to `decimal UnitPrice` in `AddItemToWaybillCommand` and `UpdateWaybillItemCommand` so the API can accept decimal UAH requests from the frontend properly.
- Updated `AddItemToWaybillHandler` and `UpdateWaybillItemHandler` to call `MoneyConversion.ToMinorUnits(request.UnitPrice)` instead of fully qualified namespace.
- Corrected test assertions across `ProzoroBanka.IntegrationTests.Api.PurchasesOperationsEndpointsTests` and `ProzoroBanka.UnitTests.Application.Purchases.Commands.AddItemToWaybill.AddItemToWaybillHandlerTests` to use the correct scale and type matching the decimal changes. Test inputs use UAH while exact Database assertions use long kopecks.

+++ Backend/src/ProzoroBanka.Application/Purchases/Commands/AddItemToWaybill/AddItemToWaybillCommand.cs
@@ -8,5 +8,5 @@ public record AddItemToWaybillCommand(
     Guid WaybillDocumentId,
     string Name,
     decimal Quantity,
-    long UnitPrice
+    decimal UnitPrice
 ) : IRequest<ServiceResponse<Guid>>;
